homeroutes removed, baseball removed, thing breaking the mysql connection commented out because Ryan and Adam may still need it for Heroku
