+++ controllers/index.js
@@ -1,9 +1,8 @@
 const router = require('express').Router();
 
 const apiRoutes = require('./api');
-const homeRoutes = require('./homeRoutes');
 
 router.use('/api', apiRoutes);
-router.use('/' , homeRoutes);
+
 
 module.exports = router;
