+++ server.js
@@ -15,9 +15,9 @@ if (process.env.NODE_ENV === 'production') {
     app.use(express.static('client/build'));
 }
 
-app.get("*", (req, res) => {
-    res.sendFile(path.resolve(__dirname, "client", "build", "index.html"));
-});
+//app.get("*", (req, res) => {
+//    res.sendFile(path.resolve(__dirname, "client", "build", "index.html"));
+//});
 
 const sess = {
     secret: 'Super secret secret',
